😱Fixing huge reentrant bug in paging.c
Basically the memory manager is called in the paging mapping function. The memory manager in turn may decide to refill it's slab allocator. This calls the paging mapping function again. If unlucky, this causes an overwrite in the tree, which leads to a later call creating and mapping the L2 page table again. Even worse, if both calls act upon the same L2 page table, then the page table is created twice at the same time, again followed by an overwrite. The result of this is that the slabs create in the refill operation become unavailable. Much disaster, such pain. This took me at least 6 hours to figure out..